+++ plugins/O-Formant/Source/PluginEditor.h
@@ -25,6 +25,9 @@ private:
     std::unique_ptr<juce::WebSliderRelay> vibratoRateRelay;
     std::unique_ptr<juce::WebSliderRelay> vibratoDepthRelay;
     std::unique_ptr<juce::WebSliderRelay> vibratoDelayRelay;
+    std::unique_ptr<juce::WebSliderRelay> jitterRelay;
+    std::unique_ptr<juce::WebSliderRelay> shimmerRelay;
+    std::unique_ptr<juce::WebSliderRelay> rdModDepthRelay;
     std::unique_ptr<juce::WebSliderRelay> consonantLevelRelay;
     std::unique_ptr<juce::WebSliderRelay> consonantToneRelay;
     std::unique_ptr<juce::WebSliderRelay> sibilanceRelay;
@@ -51,6 +54,9 @@ private:
     std::unique_ptr<juce::WebSliderParameterAttachment> vibratoRateAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> vibratoDepthAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> vibratoDelayAttachment;
+    std::unique_ptr<juce::WebSliderParameterAttachment> jitterAttachment;
+    std::unique_ptr<juce::WebSliderParameterAttachment> shimmerAttachment;
+    std::unique_ptr<juce::WebSliderParameterAttachment> rdModDepthAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> consonantLevelAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> consonantToneAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> sibilanceAttachment;